app/ModelForge-1/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
yt-dlp==2024.8.6
youtube-transcript-api==0.6.2
openai==1.40.6
sqlite-utils==3.36
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.0
google-api-python-client==2.142.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.43
trimesh==4.7.4
Werkzeug==3.1.3
itsdangerous==2.2.0
click==8.1.7
requests==2.32.3

# Enhanced 3D processing and computer vision
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0
scikit-image>=0.22.0
Pillow>=10.0.0
imageio>=2.31.0

# Video processing
ffmpeg-python>=0.2.0

# Advanced mesh operations
meshio>=5.3.0

# Performance monitoring
psutil>=5.9.0

# Optional: Deep learning for advanced features (uncomment if needed)
# torch>=2.0.0
# torchvision>=0.15.0
# transformers>=4.35.0

# Optional: Advanced dependencies (uncomment if needed)
# pymeshlab>=2023.12  # Commented out due to system dependencies not available in Docker
